+++ noita_api/Cargo.toml
@@ -11,3 +11,5 @@ shared = {path = "../shared"}
 base64 = "0.22.1"
 rustc-hash = "2.0.0"
 smallvec = "1.15.1"
+object = "0.37.1"
+rayon = "1.10.0"
