+++ noita_api/src/lib.rs
@@ -16,6 +16,7 @@ use std::{
 pub mod lua;
 pub mod serialize;
 pub use noita_api_macro::add_lua_fn;
+pub mod noita;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct EntityID(pub NonZero<isize>);
